requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
bitarray==3.0.0
build==1.2.2.post1
CacheControl==0.14.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
ckzg==2.0.1
cleo==2.1.0
crashtest==0.4.1
cytoolz==1.0.0
distlib==0.3.9
Django>=4.2.0
django-redis==5.4.0
djangorestframework>=3.14.0
dulwich==0.21.7
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==2.1.0
eth-typing==5.0.1
eth-utils==5.1.0
eth_abi==5.1.0
fastjsonschema==2.20.0
filelock==3.16.1
frozenlist==1.5.0
hexbytes==1.2.1
idna==3.10
importlib_metadata==8.5.0
installer==0.7.0
jaraco.classes==3.4.0
keyring==24.3.1
more-itertools==10.5.0
msgpack==1.1.0
multidict==6.1.0
packaging==24.2
parsimonious==0.10.0
pexpect==4.9.0
pkginfo==1.11.2
platformdirs==4.3.6
poetry-core==1.9.1
poetry-plugin-export==1.8.0
propcache==0.2.0
psycopg2-binary>=2.9.6
ptyprocess==0.7.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.1
pydantic_core==2.27.1
pyproject_hooks==1.2.0
python-decouple>=3.8
pyunormalize==16.0.0
RapidFuzz==3.10.1
redis==5.2.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rlp==4.0.1
shellingham==1.5.4
sqlparse==0.5.2
tomli==2.1.0
tomlkit==0.13.2
toolz==1.0.0
trove-classifiers==2024.10.21.16
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.27.1
web3>=6.0.0
websockets==13.1
xattr==1.1.0
yarl==1.18.0
zipp==3.21.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-mock==3.12.0
coverage==7.3.2
